+++ api/src/main/java/us/teaminceptus/novaconomy/api/events/player/economy/PlayerWithdrawEvent.java
@@ -1,4 +1,4 @@
-package us.teaminceptus.novaconomy.api.events.player;
+package us.teaminceptus.novaconomy.api.events.player.economy;
 
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
