+++ api/src/main/javadoc/us/teaminceptus/novaconomy/api/events/player/economy/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package holding of all Novaconomy API Events related to Players and their involvement in Economies
+ */
+package us.teaminceptus.novaconomy.api.events.player.economy;
